rename main logger to app-logger
